requirements.txt
pillow~=12.2.0
